requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.4.0
